requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.0.8
django-filter==2.3.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
Markdown==3.2.2
phonenumbers==8.12.7
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
PyYAML==5.3.1
pyyml==0.0.2
sqlparse==0.3.1
